create BridgeContainer
